chore(views): restructure views
add custom validation error

add entries validations

chore(auth) add auth validation
