+++ app/__init__.py
@@ -7,4 +7,8 @@ app = create_app(os.environ.get('FLASK_ENV'))
 
 
 # import views
-from .views import home
+from .views import auth, entries, profile,general_routes
+
+app.register_blueprint(auth)
+app.register_blueprint(entries)
+app.register_blueprint(profile)
