+++ app/views/__init__.py
@@ -1,0 +1,3 @@
+from .authentication import auth
+from .entries import entries
+from .profiles import profile
