+++ requirements.txt
@@ -8,3 +8,4 @@ psycopg2==2.7.5
 pyIsEmail==1.3.2
 requests>=2.20.0
 PyJWT==1.6.4
+jsonschema==3.0.1
