+++ tests/__init__.py
@@ -1,9 +1,7 @@
 """ Inintialisation file for tests"""
 from config import Testing
-from app import app
 
-app.config.from_object(Testing)
-#Test data
+# Test data
 test_user_data = {
     "username": "peter",
     "name": "Simon Peter",
@@ -19,11 +17,11 @@ test_entry = {
     "entry_content": "Test content"
 }
 wrong_test_entry = {
-    "entry_content": "Test content"
+   "entry_content": ""
 }
 wrong_test_user_data = {
     "name": "Simon Peter",
-    "email": "kimbugwe@gmail.com",
+    "email": "kimom",
     "password": "12345678"
 }
 wrong_test_sign_in = {
